Refactor: Modularize load_measurement_data to src.io
